requirements.txt
CHEM==1.3.0
FINGER==1.0
numpy==1.21.6
OpenEye_toolkits_python3_linux_x64.egg==info
rdkit==2023.3.2
rdkit_pypi==2022.9.5
sets==0.3.2
